+++ launcher/command.h
@@ -1,5 +1,9 @@
 #pragma once
-#include <webcface/webcface.h>
+#include <webcface/log.h>
+#include <webcface/func.h>
+#include <webcface/client.h>
+#include <webcface/value.h>
+#include <webcface/view.h>
 #include <spdlog/spdlog.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <process.hpp>
@@ -8,114 +12,52 @@
 #include <unordered_map>
 #include "../common/common.h"
 
-enum class CaptureMode {
-    never,
-    onerror,
-    always,
-};
 struct Process : std::enable_shared_from_this<Process> {
     std::string name;
     std::string exec;
     std::string workdir;
-    CaptureMode capture_stdout;
+    bool capture_stdout;
     bool stdout_is_utf8;
     std::unordered_map<std::string, std::string> env;
     int exit_status = 0;
     std::shared_ptr<TinyProcessLib::Process> p;
+    bool running_prev = false;
     std::shared_ptr<spdlog::logger> logger;
-    std::string logs;
+    std::optional<webcface::Log> send_logs = std::nullopt;
+    std::string logs_out, logs_err;
 
     Process(const Process &) = delete;
     Process &operator=(const Process &) = delete;
     Process(const std::string &name, const std::string &exec,
-            const std::string &workdir, CaptureMode capture_stdout,
+            const std::string &workdir, bool capture_stdout,
             bool stdout_is_utf8,
-            const std::unordered_map<std::string, std::string> &env)
-        : std::enable_shared_from_this<Process>(), name(name), exec(exec),
-          workdir(workdir), capture_stdout(capture_stdout),
-          stdout_is_utf8(stdout_is_utf8), env(env),
-          logger(spdlog::stdout_color_mt(name)) {
-        logger->set_pattern("[%n] %v");
-    }
+            const std::unordered_map<std::string, std::string> &env);
 
-    void start() {
-        auto read_log = [cmd = shared_from_this()](const char *bytes,
-                                                   std::size_t n) {
-#ifdef _WIN32
-            if (!cmd->stdout_is_utf8) {
-                cmd->logs += acpToUTF8(bytes, static_cast<int>(n));
-            } else {
-                cmd->logs.append(bytes, n);
-            }
-#else
-            cmd->logs.append(bytes, n);
-#endif
-            cmd->logger->info(std::string(bytes, n));
-        };
-        if (is_running()) {
-            spdlog::warn("Command '{}' is already started.", this->name);
-            throw std::runtime_error("already started");
-        } else {
-            spdlog::info("Starting command '{}'.", this->name);
-            this->logs.clear();
-            if (this->capture_stdout != CaptureMode::never) {
-                p = std::make_shared<TinyProcessLib::Process>(
-                    this->exec, this->workdir, this->env, read_log, read_log);
-            } else {
-                p = std::make_shared<TinyProcessLib::Process>(
-                    this->exec, this->workdir, this->env);
-            }
-        }
-    }
-    void kill([[maybe_unused]] int sig) {
-        if (is_running()) {
-#ifdef _WIN32
-            spdlog::info("Stopping command '{}'.", this->name);
-            p->kill(false);
-#else
-            spdlog::info("Sending signal {} to command '{}'.", sig, this->name);
-            p->signal(sig);
-#endif
-        } else {
-            spdlog::warn("Command '{}' is already stopped.", this->name);
-            throw std::runtime_error("already stopped");
-        }
-    }
+    void readLog(const char *bytes, std::size_t n, bool is_stderr);
+    void start();
+    void kill([[maybe_unused]] int sig);
 
-    bool is_running() { return p && !p->try_get_exit_status(exit_status); }
+    bool is_running();
 };
 
 // ProcessにStart/Stopボタンの実装を追加したもの
 struct Command : std::enable_shared_from_this<Command> {
-    webcface::Func start_f, stop_f;
+    webcface::Func start_f, stop_f, kill_f;
+    std::vector<webcface::CallHandle> stop_h;
     std::shared_ptr<Process> start_p;
-    using StopOption = std::variant<std::nullopt_t, std::shared_ptr<Process>, int>;
-    StopOption stop_p;
+    using StopOption =
+        std::variant<std::nullopt_t, std::shared_ptr<Process>, int>;
+    StopOption stop_p, kill_p;
 
     Command(const Command &) = delete;
     Command &operator=(const Command &) = delete;
-    Command(const std::shared_ptr<Process> &start_p,
-            const StopOption &stop_p)
-        : start_p(start_p), stop_p(stop_p) {}
+    Command(const std::shared_ptr<Process> &start_p, const StopOption &stop_p,
+            const StopOption &kill_p)
+        : start_p(start_p), stop_p(stop_p), kill_p(kill_p) {}
 
     // shared_from_thisを使うためコンストラクタと別
-    void initFunc(WebCFace::Client &wcli) {
-        start_f =
-            wcli.func(start_p->name + "/start").set([cmd = shared_from_this()] {
-                cmd->start_p->start();
-            });
-        stop_f =
-            wcli.func(start_p->name + "/stop").set([cmd = shared_from_this()] {
-                switch(cmd->stop_p.index()){
-                case 1:
-                    std::get<1>(cmd->stop_p)->start();
-                    break;
-                case 2:
-                    cmd->start_p->kill(std::get<2>(cmd->stop_p));
-                    break;
-                default:
-                    throw std::runtime_error("stop signal disabled");
-                }
-            });
-    }
+    void initFunc(webcface::Client &wcli);
+
+    void update(webcface::Client &wcli);
+    void updateView(webcface::View &v);
 };